fix(icon): move icon component to client directory due to use of as prop
