+++ src/components/client/core/Drawer/Drawer.test.tsx
@@ -13,9 +13,10 @@ import {
   DrawerDescription,
   DrawerCloseTrigger,
   DrawerTrigger,
+  Icon,
   Portal,
 } from "components/client";
-import { Icon, Text } from "components/universal";
+import { Text } from "components/universal";
 
 const DrawerTest = () => (
   <Drawer>
