+++ src/components/universal/core/index.ts
@@ -1,5 +1,4 @@
 export { default as Card, CardHeader, CardBody, CardFooter } from "./Card/Card";
-export { default as Icon } from "./Icon/Icon";
 export { default as Text } from "./Text/Text";
 
 export type {
@@ -8,5 +7,4 @@ export type {
   CardBodyProps,
   CardFooterProps,
 } from "./Card/Card";
-export type { Props as IconProps } from "./Icon/Icon";
 export type { Props as TextProps } from "./Text/Text";
